moved waf into extract.py, fixed tests, added make down/clean
